Update README.md
Our project is a voluntary organization that specializes in donating clothing to families in need. We called it the Hope Organization.
We built this site using the HTML, CSS bootstrap webform in asp.net C#, our website contains two roles, the volunteer and the permanent donor in addition to the admin.